Relocate common setup - vars, flags, init()
Move any setup that is common to all tests - which includes setting global
variables, parsing CLI flags, and doing the pre-flight health check - to a
separate file called `common_setup_test.go` so that it's logically
separated.

The naming of this file seems to have no effect on the setup being performed
before running any of the tests. Presumably because it compiles all test
files that belong to the "main" package.
