+++ cdn_acceptance_test.go
@@ -1,50 +1,12 @@
 package main
 
 import (
-	"crypto/tls"
-	"flag"
 	"fmt"
 	"io/ioutil"
-	"log"
 	"net"
 	"net/http"
 	"testing"
-	"time"
 )
-
-const requestTimeout = time.Second * 5
-
-var (
-	edgeHost      = flag.String("edgeHost", "www.gov.uk", "Hostname of edge")
-	originPort    = flag.Int("originPort", 8080, "Origin port to listen on for requests")
-	skipVerifyTLS = flag.Bool("skipVerifyTLS", false, "Skip TLS cert verification if set")
-
-	client       *http.Transport
-	originServer *CDNServeMux
-)
-
-// Setup clients and servers.
-func init() {
-
-	flag.Parse()
-
-	tlsOptions := &tls.Config{}
-	if *skipVerifyTLS {
-		tlsOptions.InsecureSkipVerify = true
-	}
-
-	client = &http.Transport{
-		ResponseHeaderTimeout: requestTimeout,
-		TLSClientConfig:       tlsOptions,
-	}
-	originServer = StartServer(*originPort)
-
-	log.Println("Confirming that CDN is healthy")
-	err := confirmEdgeIsHealthy(originServer, *edgeHost)
-	if err != nil {
-		log.Fatal(err)
-	}
-}
 
 func TestHelpers(t *testing.T) {
 	testHelpersCDNServeMuxHandlers(t, originServer)
